requirements.txt
psycopg2-binary>=2.9
requests>=2.28
pandas>=1.5
python-dotenv>=1.0